Split controller.go off of the ClusterIngress reconciler
